test(home-screen): move backup nudge focus gating into the real HomeScreen spec
